+++ IsThereAnyNews.DataAccess/Implementation/UserRepository.cs
@@ -72,14 +72,14 @@ namespace IsThereAnyNews.DataAccess.Implementation
             return userPublicProfiles.Distinct().ToList();
         }
 
-        public User LoadUserPublicProfile(int id)
+        public User LoadUserPublicProfile(long id)
         {
             var user = this.database.Users
                 .Include(x => x.RssSubscriptionList)
                 .Include(x => x.RssSubscriptionList.Select(c => c.RssChannel))
                 .Include(x => x.EventsRssViewed)
-                .Include(x => x.EventsRssViewed.Select(e=>e.RssEntry))
-                .Single(x => x.Id==id);
+                .Include(x => x.EventsRssViewed.Select(e => e.RssEntry))
+                .Single(x => x.Id == id);
             return user;
         }
     }
