+++ IsThereAnyNews.Services/IUsersService.cs
@@ -5,6 +5,6 @@ namespace IsThereAnyNews.Services
     public interface IUsersService
     {
         AllUsersPublicProfilesViewModel LoadAllUsersPublicProfile();
-        UserDetailedPublicProfileViewModel LoadUserPublicProfile(int id);
+        UserDetailedPublicProfileViewModel LoadUserPublicProfile(long id);
     }
 }
